requirements.txt
fastapi
opencensus
opencensus-ext-azure
opencensus-ext-logging
uvicorn
python-dotenv
pytest